RC-103 - Minor fixes
